chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-unsafe_inline-expected.txt

This is a testharness.js-based test.
PASS 'strict-dynamic' is ineffective for `style-src`.
PASS 'unsafe-inline' is properly subsumed in `style-src`.
PASS 'unsafe-inline' is only ineffective if the effective returned csp has nonces in `style-src`.
PASS 'unsafe-inline' is only ineffective if the effective returned csp has hashes in `style-src`.
PASS Returned csp does not have to allow 'unsafe-inline' in `style-src` to be subsumed.
PASS 'unsafe-inline' does not matter if returned csp is effectively `none`.
PASS 'unsafe-inline' is properly subsumed in `script-src`.
PASS Returned csp only loads 'unsafe-inline' scripts with 'nonce-abc'.
PASS 'unsafe-inline' is ineffective when nonces are present.
PASS 'unsafe-inline' is only ineffective if the effective returned csp has hashes in `script-src`.
FAIL Required csp allows `strict-dynamic`, but retuned csp does. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Required csp does not allow `unsafe-inline`, but retuned csp does. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Returned csp allows a nonce. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Returned csp allows a hash. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Effective returned csp allows 'unsafe-inline' assert_unreached: No message should be sent from the frame. Reached unreachable code
PASS Effective returned csp does not allow 'sha512-321cba' hash.
Harness: the test ran to completion.
